Suggested change

+++ Source/Components/Loyalty/Broadcaster/MLBusinessLoyaltyBroadcastData.swift
@@ -9,7 +9,7 @@
 public class MLBusinessLoyaltyBroadcastData: NSObject {
 
     public let level: Int
-    public let percentage: Double
+    public let percentage: Float
     public let primaryColor: String
 
     public init(level: Int, percentage: Double, primaryColor: String) {
